+++ src/frontend/app/events/_featureStyleDefinitions.js
