+++ src/frontend/app/events/forms/Delays.jsx
@@ -2,9 +2,9 @@ import { useCallback, useState } from 'react';
 
 import Select from 'react-select';
 
-import { TrafficImpacts } from './references';
-import { DraggableRows } from './shared';
-import { selectStyle } from '../components/Map/helpers';
+import { TrafficImpacts } from '../references';
+import { DraggableRows } from '../shared';
+import { selectStyle } from '../../components/Map/helpers';
 
 export default function Delays({ event, dispatch }) {
   return (
