+++ src/frontend/app/events/forms/Details.jsx
@@ -2,8 +2,8 @@ import { useState, useRef } from 'react';
 
 import Select from 'react-select';
 
-import { FORM_CATEGORIES, FORM_CATEGORY_PHRASE, FORM_PHRASES, FORM_PHRASE_CATEGORY, PHRASES_LOOKUP } from "./references";
-import { selectStyle } from '../components/Map/helpers';
+import { FORM_CATEGORIES, FORM_CATEGORY_PHRASE, FORM_PHRASES, FORM_PHRASE_CATEGORY, PHRASES_LOOKUP } from "../references";
+import { selectStyle } from '../../components/Map/helpers';
 
 export default function Details({ errors, event, dispatch }) {
   const catRef = useRef();
