+++ src/frontend/app/slices/visibleLayers.js
@@ -3,11 +3,13 @@ import { createSlice } from '@reduxjs/toolkit';
 import { defaultLayers } from '../events/Layers';
 
 const layersInitial = () => {
-  try {
-    const stored = localStorage.getItem('visibleLayers');
-    if (stored) { return JSON.parse(stored); }
-  } catch (err) {
-    console.log(err);
+  if (globalThis.localStorage) {
+    try {
+      const stored = localStorage.getItem('visibleLayers');
+      if (stored) { return JSON.parse(stored); }
+    } catch (err) {
+      console.log(err);
+    }
   }
   return defaultLayers;
 }
